Aggregate consecutive pixels per spot

Helper function for estimating the spillover matrix. Per metal spot, consecutive pixels a aggregated (default: summed).

binAcrossPixels(
  object,
  bin_size,
  spot_id = "sample_id",
  assay_type = "counts",
  statistic = "sum",
  ...
)

Arguments

object
a SingleCellExperiment object containing pixel intensities per channel. Individual pixels are stored as columns and channels are stored as rows.

bin_size
single numeric indicating how many consecutive pixels should be aggregated.

spot_id
character string indicating which colData(object) entry stores the isotope names of the spotted metal.

assay_type
character string indicating which assay to use.

statistic
character string indicating the statistic to use for aggregating consecutive pixels.

...
additional arguments passed to aggregateAcrossCells

Value

returns a SCE object

Examples