Generates an SCE data container for single-cell IMC data generated by ImcSegmentationPipeline (https://github.com/BodenmillerGroup/ImcSegmentationPipeline)

This function takes various inputs and stores it in an SingleCellExperiment data container

generateSCE(
  cells = NULL,
  cell_meta = c("ImageNumber", "ObjectNumber"),
  image_meta = NULL,
  panel_meta = NULL,
  metal_tag = "Metal.Tag",
  channel_number = "channel",
  target_names = NULL,
  imageNumber = "ImageNumber",
  objectNumber = "ObjectNumber",
  measurement_column = NULL,
  scaling_column = NULL,
  metadata = NULL
)

Arguments

cells
data.frame containing single-cell measurements (cell.csv)

cell_meta
character vector containing column names of relevant meta data stored in `cells`, must at least include a sort of `CellNumber` and `ImageNumber`

image_meta
data.frame containing meta data for the acquired samples (image.csv)

panel_meta
data.frame containing meta data for the applied panel

metal_tag
colnames of metal tag column in `panel_meta` (default = "Metal.Tag")

channel_number
colnames of column that contains the channel number in `panel_meta` (default = "channel")

target_names
colnames of column that contains the antibody/target name in `panel_meta`

imageNumber
name of column containing an sample identifier present in all provided input

objectNumber
name of column containing an object number for individual cells

measurement_column
shared colnames pattern of cells columns storing channel measurement (e.g. Intensity_MeanIntensityCorrected_FullStackFiltered)

scaling_column
colnames of scaling column storing the applied scaling factor (e.g. Scaling_FullStack)

metadata
(optional) data.frame containing further metadata associated to cells, must include `imageID`

Value

returns SCE object

Examples

#sce <- generateSCE(cells = cells,
#       cell_meta = c("ImageNumber", "ObjectNumber"),
#       image_meta = image_mat,
#       panel_meta = panel_mat,
#       metal_tag = "Metal.Tag",
#       channel_number = "channel", 
#       target_names = "clean_target",
#       imageNumber = "ImageNumber",
#       objectNumber = "ObjectNumber",
#       measurement_column = "Intensity_MeanIntensityComp_FullStackFiltered",
#       scaling_column = "Scaling_FullStack")