Generates a SingleCellExperiment from .txt files

Helper function to process raw .txt files acquired by the Hyperion system into a SingleCellExperiment container. This function is mainly used to read-in data generated from a "spillover slide". Here, each .txt file contains the measurements of multiple pixels for a single stain across all open channels.

prepareSCEfromTXT(
  txt_list,
  metadata_cols = c("Start_push", "End_push", "Pushes_duration", "X", "Y", "Z"),
  verbose = TRUE
)

Arguments

txt_list
a named list containing the read-in txt files for each spot. Entries to txt_list need to be coercible to data.frame objects.

metadata_cols
character vector indicating which column entries of the .txt files should be saved in the colData(sce) slot.

verbose
logical indicating if additional information regarding the spotted and acquired masses should be shown.

Value

returns a SCE object

Examples