Display all features measured by CellProfiler.

Searchable datatable object of cell and image features as extracted by CellProfiler.

show_cpout_features(
  path,
  display = c("cell_features", "image_features"),
  cell_features = "var_cell.csv",
  image_features = "var_Image.csv"
)

Arguments

path

full path to the CellProfiler output folder

display

single character indicating which features to display. Accepted entries are cell_features to display extracted single-cell features or image_features to display extracted image-level features.

cell_features

single character indicating the name of the file storing the extracted cell features.

image_features

single character indicating the name of the file storing the extracted image features.

Value

a datatable object

See also

Author

Nils Eling (nils.eling@dqbm.uzh.ch)

Examples

path <- system.file("extdata/mockData/cpout", package = "imcRtools")

# Display cell features 
show_cpout_features(path)

# Display image features 
show_cpout_features(path, display = "image_features")